Move docker reference logic to reference/docker package
